TP-202 결제 서버 분리

+++ src/main/java/com/goorm/tablepick/domain/reservation/dto/response/ReservationResponseDto.java
@@ -19,7 +19,7 @@ public class ReservationResponseDto {
     private Long id;
 
     @Schema(description = "예약 인원 수", example = "3")
-    private Long partySize;
+    private int partySize;
 
     @Schema(description = "예약 날짜", example = "2025-05-08")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
